Refactor: Consolidate semantic analysis modules
- Merged `control_flow.rs`, `declarations.rs`, and `patterns.rs` into `statements.rs` to flatten module hierarchy and centralize statement parsing logic.
- Moved `detect_iterator_pattern` to `conversion.rs` to align with expression conversion responsibility.
- Moved `resolve_type_name` to `types.rs` for better cohesion.
- Simplified `src/semantic` module structure.
