+++ src/semantic/mod.rs
@@ -32,13 +32,11 @@
 //! without changing the semantic meaning.
 
 pub mod assembler;
-pub(crate) mod control_flow;
 pub(crate) mod conversion;
-pub(crate) mod declarations;
 pub mod expressions;
 pub(crate) mod model;
-pub(crate) mod patterns;
 mod resolver;
+pub(crate) mod statements;
 mod types;
 
 pub use crate::morphology::{DisambiguationContext, analyze_article, disambiguate, resolve_best};
@@ -52,13 +50,12 @@ pub use types::*;
 use crate::ast::{Expr, Program, Statement};
 use crate::errors::GlossaError;
 
-use self::control_flow::analyze_control_flow;
 use self::conversion::convert_assembled_to_analyzed;
-use self::declarations::{
-    analyze_test_declaration, analyze_trait_definition, analyze_trait_impl, analyze_type_definition,
-};
 use self::expressions::feed_expr_to_assembler_with_context;
-use self::patterns::{try_parse_struct_instantiation, try_parse_trait_method_call};
+use self::statements::{
+    analyze_control_flow, analyze_test_declaration, analyze_trait_definition, analyze_trait_impl,
+    analyze_type_definition,
+};
 
 /// Perform semantic analysis on a program using the slot-based assembler
 /// This is the primary entry point that provides word-order independence
@@ -260,7 +257,8 @@ pub fn analyze_statement(
     stmt: &Statement,
     scope: &mut Scope,
 ) -> Result<Vec<AnalyzedStatement>, GlossaError> {
-    // Check for control flow (if, while, etc.)
+    // Check for control flow (if, while, etc.) AND patterns (struct inst, trait calls)
+    // The consolidated logic in `statements.rs` handles all of these.
     if let Some(control_flow) = analyze_control_flow(stmt, scope)? {
         // If it's a function definition, register it in the scope
         if let AnalyzedStatement::FunctionDef {
@@ -278,16 +276,6 @@ pub fn analyze_statement(
         }
 
         return Ok(vec![control_flow]);
-    }
-
-    // Check for struct instantiation pattern
-    if let Some(struct_inst) = try_parse_struct_instantiation(stmt, scope)? {
-        return Ok(vec![struct_inst]);
-    }
-
-    // Check for trait method call pattern
-    if let Some(method_call) = try_parse_trait_method_call(stmt, scope)? {
-        return Ok(vec![method_call]);
     }
 
     // Check if it's a block statement (regular statement containing a single block expression)
